# src/main/resources/doc/openapi.json
{
  "components" : {
    "schemas" : {
...
      "edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "error", "message", "path", "status", "timestamp" ],
        "title" : "ApiErrorResponse",
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "REST API of Ultimatum Game. Generated from source annotations.",
    "title" : "Ultimatum Game API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/statistics/{sessionId}/csv" : {
      "get" : {
        "operationId" : "downloadCsv",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "statistic-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}